Revert "EUIMessages is enum class now"
This reverts commit 23a7cf25faae544c24fa1572f281f7c2b02ce68d.

+++ src/xrGame/ui/UITabControl.h
@@ -42,7 +42,7 @@ public:
 	const shared_str&	GetPrevActiveId				()								{ return m_sPrevPushedId; }
 			void		SetActiveTab				(const shared_str& sNewTab);
 			void		SetActiveTab_script			(LPCSTR sNewTab)				{SetActiveTab(sNewTab);};
-	const	u32			GetTabsCount				() const						{ return (u32)m_TabsArr.size(); }
+	const	u32			GetTabsCount				() const						{ return m_TabsArr.size(); }
 	
 	// –ежим клавилатурных акселераторов (вкл/выкл)
 	IC bool				GetAcceleratorsMode			() const						{ return m_bAcceleratorsEnable; }
